feat: Add packaging utility to DETR.
I needed DETR for another model. I found it more convenient to package the whole repo.
The tests are passing and I can run the notebooks without workaround, so it seems to work so far.
Basically just moved modules and renamed import in package format.
See README.md for update on the changes.